+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.3 (Nov 6, 2023)
+
+- Added `davidhirtz\yii2\media\models\traits\AssetParentTrait` and added support for array values in HTML sizes
+  attribute
+- Added `Module::$breakpoints` for the HTML sizes attribute
+- Renamed `getSrcsetSizes()` to `getSizes()`
+
 ## 2.0.2 (Nov 6, 2023)
 
 - Added `File::isAudio()` and `File::isVideo()`